Started on refactoring and moving code. No finished commit, just making sure the code is backed up
