+++ pyplanet/apps/contrib/nightcup/templates/ncstandings.xml
@@ -3,28 +3,25 @@
 {% block content %}
   {% if cps != None %}
     {% for time in cps %}
-      <frame pos= {% if (time.nr_qualified - 2) == loop.index0 %}
-                "0 -{{ (loop.index0 * 3.25) }} - 5"
-                {% else %} "0 -{{ (loop.index0 * 3.25) }}"
-                {%endif %}>
+      <frame pos="0 -{{ (loop.index0 * 3.25) }}">
         <quad pos="0 0" z-index="0" size="4 3" bgcolor=
-        {% if time.virt_qualified == true %} "32CD3260"
+        {% if time.virt_qualified %} "32CD3260"
         {% elif time.virt_eliminated %} "DC143C60"
         {% else %} "00000070" {%endif %}/>
-        <label pos="2 -1.5" z-index="1" size="4 3" text="{{ time.cp }}" textsize="0.3" textfont="RajdhaniMono"
+        <label pos="2 -1.5" z-index="1" size="4 3" text="{{ time.col0 }}" textsize="0.3" textfont="RajdhaniMono"
                textemboss="1" halign="center" valign="center2"/>
         <quad pos="4.25 0" z-index="0" size="20.5 3" bgcolor=
-        {% if time.virt_qualified == true %} "32CD3260"
+        {% if time.virt_qualified %} "32CD3260"
         {% elif time.virt_eliminated %} "DC143C60"
         {% else %} "00000070" {%endif %}/>
         <label pos="4.5 -1.5" z-index="1" size="20 3" text="{{ time.nickname }}" textsize="0.2" textfont="RajdhaniMono"
                textemboss="1" halign="left" valign="center2" style="TextCardSmall"
                action="pyplanet__widgets_nightcupstandings__spec_{{ time.login }}"/>
         <quad pos="25 0" z-index="0" size="10 3" bgcolor=
-        {% if time.virt_qualified == true %} "32CD3260"
+        {% if time.virt_qualified %} "32CD3260"
         {% elif time.virt_eliminated %} "DC143C60"
         {% else %} "00000070" {%endif %}/>
-        <label pos="30 -1.5" z-index="1" size="10 3" text="{{ time.color }}{{ time.cptime }}" textsize="0.2"
+        <label pos="30 -1.5" z-index="1" size="10 3" text="{{ time.color }}{{ time.col2 }}" textsize="0.2"
                textfont="RajdhaniMono" textemboss="1" halign="center" valign="center2"/>
       </frame>
     {% endfor %}
